+++ examples/demo-rollup/tests/prover/mod.rs
@@ -1,4 +1,3 @@
-use std::path::PathBuf;
 use std::sync::atomic::AtomicU64;
 use std::sync::Arc;
 
@@ -9,18 +8,14 @@ use sov_db::storage_manager::NativeStorageManager;
 use sov_mock_da::{MockAddress, MockBlock, MockDaService, MockDaSpec};
 use sov_mock_zkvm::MockZkvm;
 use sov_modules_api::execution_mode::WitnessGeneration;
-use sov_modules_api::{OperatingMode, SlotData, Spec, ZkVerifier};
+use sov_modules_api::{OperatingMode, SlotData};
 use sov_modules_stf_blueprint::{GenesisParams, StfBlueprint};
 use sov_rollup_interface::da::BlockHeaderTrait;
 use sov_rollup_interface::node::da::DaService;
 use sov_rollup_interface::stf::{ExecutionContext, StateTransitionFunction};
 use sov_rollup_interface::storage::HierarchicalStorageManager;
-use sov_rollup_interface::zk::{
-    StateTransitionPublicData, StateTransitionWitness, StateTransitionWitnessWithAddress, ZkvmHost,
-};
-use sov_sp1_adapter::host::SP1Host;
-use sov_sp1_adapter::BlockHeaderWithProof;
-use sov_sp1_adapter::{SP1MethodId, SP1Verifier, SP1};
+use sov_rollup_interface::zk::StateTransitionWitness;
+use sov_sp1_adapter::SP1;
 use sov_state::ProverStorage;
 use sov_test_utils::generators::BlobBuildingCtx;
 use sov_test_utils::TestStorageSpec;
@@ -40,6 +35,7 @@ pub(super) type DefaultSpec = sov_modules_api::configurable_spec::ConfigurableSp
 
 mod datagen;
 mod network;
+mod sp1_cpu_prover;
 
 pub(super) type TestSTF = StfBlueprint<DefaultSpec, Runtime<DefaultSpec>>;
 pub(super) type ProofStateRoot =
@@ -47,42 +43,6 @@ pub(super) type ProofStateRoot =
 pub(super) type ProofWitness =
     <TestSTF as StateTransitionFunction<SP1, MockZkvm, MockDaSpec>>::Witness;
 pub(super) type StfWitness = StateTransitionWitness<ProofStateRoot, ProofWitness, MockDaSpec>;
-
-type ProofInput = StateTransitionWitnessWithAddress<
-    <DefaultSpec as Spec>::Address,
-    ProofStateRoot,
-    ProofWitness,
-    MockDaSpec,
->;
-
-#[tokio::test(flavor = "multi_thread")]
-#[ignore = "This test is used to generate data for testing the aggregate proof circuit and should be enabled only when needed."]
-async fn test_save_proofs() {
-    let host = TestHost::new().await;
-    let proof_data = generate_proofs(true, &host).await;
-    let proofs_dir = PathBuf::from(env!("CARGO_MANIFEST_DIR"))
-        .join("tests")
-        .join("test_data")
-        .join("tmp");
-
-    std::fs::create_dir_all(&proofs_dir).unwrap();
-    std::fs::write(proofs_dir.join("inner_vk.bin"), host.verifying_key_bytes()).unwrap();
-
-    for (i, data) in proof_data.into_iter().enumerate() {
-        let proof_public_data = host.verify(data.proof.clone()).await;
-        assert_eq!(proof_public_data.slot_hash, data.da_block_header.hash());
-        let json = serde_json::to_string(&data).unwrap();
-        std::fs::write(proofs_dir.join(format!("inner_{i}_proof.json")), &json).unwrap();
-    }
-}
-
-/// This test reproduces the proof generation process for the rollup used in benchmarks.
-#[tokio::test(flavor = "multi_thread")]
-#[cfg_attr(skip_guest_build, ignore)]
-async fn test_proof_generation() {
-    let host = TestHost::new().await;
-    let _ = generate_proofs(false, &host).await;
-}
 
 /// Executes the STF against mock DA blocks and produces per-block witnesses.
 ///
@@ -175,86 +135,4 @@ pub(super) async fn generate_witnesses() -> (ProofStateRoot, Vec<StfWitness>) {
     }
 
     (genesis_state_root, witnesses)
-}
-
-async fn generate_proofs(
-    with_proof: bool,
-    host: &TestHost,
-) -> Vec<BlockHeaderWithProof<MockDaSpec>> {
-    let (_genesis_state_root, witnesses) = generate_witnesses().await;
-
-    let prover_address = <DefaultSpec as Spec>::Address::try_from([0u8; 28].as_ref()).unwrap();
-    let mut proofs = Vec::new();
-
-    for witness in witnesses {
-        let da_block_header = witness.da_block_header.clone();
-
-        let data: ProofInput = StateTransitionWitnessWithAddress {
-            stf_witness: witness,
-            prover_address,
-        };
-
-        let proof = host.run(data, with_proof).await;
-        proofs.push(BlockHeaderWithProof {
-            da_block_header,
-            proof,
-        });
-    }
-
-    proofs
-}
-
-// The SP1 prover manages its own Tokio runtime, which conflicts with the `tokio::test` runtime.
-// To avoid this, all blocking work must be executed inside `tokio::task::spawn_blocking`.
-struct TestHost {
-    host: SP1Host<'static>,
-    code_commitment: SP1MethodId,
-}
-
-impl TestHost {
-    async fn new() -> Self {
-        let host = SP1Host::new(*sp1::SP1_GUEST_MOCK_ELF);
-        let host_clone = host.clone();
-        let code_commitment = tokio::task::spawn_blocking(move || -> SP1MethodId {
-            host_clone
-                .code_commitment()
-                .expect("SP1 code commitment should be created successfully")
-        })
-        .await
-        .unwrap();
-
-        Self {
-            host,
-            code_commitment,
-        }
-    }
-
-    async fn run(&self, data: ProofInput, with_proof: bool) -> Vec<u8> {
-        let mut host = self.host.clone();
-        tokio::task::spawn_blocking(move || -> Vec<u8> {
-            host.add_hint(data);
-            host.run(with_proof)
-                .expect("Prover should run successfully")
-        })
-        .await
-        .unwrap()
-    }
-
-    fn verifying_key_bytes(&self) -> Vec<u8> {
-        self.code_commitment.0.clone()
-    }
-
-    #[allow(dead_code)]
-    async fn verify(
-        &self,
-        proof: Vec<u8>,
-    ) -> StateTransitionPublicData<<DefaultSpec as Spec>::Address, MockDaSpec, ProofStateRoot> {
-        let code_commitment = self.code_commitment.clone();
-        tokio::task::spawn_blocking(move || -> StateTransitionPublicData<<DefaultSpec as Spec>::Address, MockDaSpec, ProofStateRoot> {
-            SP1Verifier::verify(&proof, &code_commitment)
-                .expect("SP1 proof verification should succeed")
-        })
-        .await
-        .unwrap()
-    }
 }
